Fix scaling issue for the file and save dialogs
Fixes #1935

Our `file-dialog` and `save-dialog` have scaling issues
when the screen's viewport is limited. Since our dialogs
are not implemented using React (which would handle scaling much easier),
the problem was addressed using css media queries. This means that when
the view becomes limited at 700px, the size of the `file-tree` is
reduced allowing the entire dialog to more easily visible and the
actions buttons still accessible.

Before, the buttons were not visible which meant the useability
of the dialog was reduced.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>